updated user specific routes to check for session data

+++ flask_app/flask06.py
@@ -85,36 +85,41 @@ def register():
 
 @app.route('/notes/delete/<note_id>', methods=['POST'])
 def delete_note(note_id):
-    my_note = db.session.query(Note).filter_by(id=note_id).one()
-    db.session.delete(my_note)
-    db.session.commit()
+    if session.get('user'):
+        my_note = db.session.query(Note).filter_by(id=note_id).one()
+        db.session.delete(my_note)
+        db.session.commit()
 
-    return redirect(url_for('get_notes'))
+        return redirect(url_for('get_notes'))
+    else:
+        return redirect(url_for('login'))
 
 
 
 @app.route('/notes/edit/<note_id>', methods=['GET', 'POST'])
 def update_note(note_id):
+    if session.get('user'):
+        if request.method == 'POST':
+            title = request.form['title']
 
-    if request.method == 'POST':
-        title = request.form['title']
+            text = request.form['noteText']
 
-        text = request.form['noteText']
+            note = db.session.query(Note).filter_by(id=note_id).one()
 
-        note = db.session.query(Note).filter_by(id=note_id).one()
+            note.title = title
+            note.text = text
 
-        note.title = title
-        note.text = text
+            db.session.add(note)
+            db.session.commit()
 
-        db.session.add(note)
-        db.session.commit()
+            return redirect(url_for('get_notes'))
+        else:
+            a_user = db.session.query(User).filter_by(email='jmill315@uncc.edu').one()
+            my_note = db.session.query(Note).filter_by(id=note_id).one()
 
-        return redirect(url_for('get_notes'))
+            return render_template('new.html', note=my_note, user=session('user'))
     else:
-        a_user = db.session.query(User).filter_by(email='jmill315@uncc.edu').one()
-        my_note = db.session.query(Note).filter_by(id=note_id).one()
-
-        return render_template('new.html', note=my_note, user=a_user)
+        return redirect(url_for('login'))
 
 
 @app.route('/notes/new', methods=['GET', 'POST'])
@@ -122,20 +127,21 @@ def new_note():
 
     # a_user = {'name' : 'Jonathan', 'email' : 'jmill315@uncc.edu'}
 
-    print('request method is', request.method)
-    if request.method == 'POST':
-        title = request.form['title']
-        text = request.form['noteText']
-        from datetime import date
-        today = date.today()
-        today = today.strftime("%m-%d-%Y")
-        new_record = Note(title, text, today)
-        db.session.add(new_record)
-        db.session.commit()
-        return redirect(url_for('get_notes'))
+    if session.get('user'):
+        if request.method == 'POST':
+            title = request.form['title']
+            text = request.form['noteText']
+            from datetime import date
+            today = date.today()
+            today = today.strftime("%m-%d-%Y")
+            new_record = Note(title, text, today, session['user_id'])
+            db.session.add(new_record)
+            db.session.commit()
+            return redirect(url_for('get_notes'))
+        else:
+            return render_template('new.html', user = session['user'])
     else:
-        a_user = db.session.query(User).filter_by(email='jmill315@uncc.edu').one()
-        return render_template('new.html', user = a_user)
+        return redirect(url_for('login'))
 @app.route('/notes/<note_id>')
 def get_note(note_id):
     # notes = {1:{'title' : 'First note', 'text' : 'This is my first note', 'date' : '10-1-2020'},
@@ -161,8 +167,9 @@ def get_notes():
 @app.route('/')
 @app.route('/index')
 def index():
-    a_user = db.session.query(User).filter_by(email='jmill315@uncc.edu').one()
-    return render_template('index.html', user=a_user)
+    if session.get('user'):
+        return render_template('index.html', user = session['user'])
+    return render_template("index.html")
 
 
 app.run(host=os.getenv('IP', '127.0.0.1'),port=int(os.getenv('PORT', 5000)),debug=True)
